refactor: extract and test diff-runtime-deps closure logic
Move the IO-free dependency-closure and comparison logic out of the
`diff-runtime-deps.ts` bin into `src/diff-runtime-deps.ts` so it can be
unit-tested, and add a vitest suite covering `parseTrackedDepNames`
(allowlist/blocklist/precedence) and a Vite-bump closure diff. The bin
now keeps only CLI parsing, git/fs reads, and rendering.

Also extend the base-ref classification borrowing to `trackedDevDependencies`
(it previously only handled `untrackedDevDependencies`), so introducing the
field on a branch no longer reports every tracked devDependency as "added"
when the versions are unchanged.
